requirements.txt
# Auto-PPT MCP stack (Python 3.10+)
mcp[cli]>=1.2.0
python-pptx>=1.0.2
python-dotenv>=1.0.0
huggingface_hub>=0.26.0
Pillow>=10.0.0
duckduckgo-search>=6.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0.3
python-multipart>=0.0.9
httpx>=0.27.0
